feat(claim): add claim button to history
