+++ src/pages/Shuttle/ClaimModal/ShuttleButton/index.js
@@ -1,3 +1,2 @@
 export {default as ShuttleInButton} from './In'
 export {default as ShuttleOutButton} from './Out'
-export {default as ShuttleClaimButton} from './Claim'
