+++ src/pages/Shuttle/index.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 import {useState, useEffect} from 'react'
 import queryString from 'query-string'
 import {useHistory, useLocation} from 'react-router-dom'
@@ -15,14 +14,9 @@ import {
   KeyOfCfx,
   KeyOfBtc,
 } from '../../constants/chainConfig'
-import {
-  TxReceiptModalType,
-  TypeAccountStatus,
-  ShuttleStatus,
-} from '../../constants'
+import {TypeAccountStatus} from '../../constants'
 import ConfirmModal from './ConfirmModal'
 import ClaimModal from './ClaimModal'
-import {TransactionReceiptionModal} from '../components'
 import {useShuttleState} from '../../state'
 import {getChainIdRight} from '../../utils'
 
@@ -34,8 +28,7 @@ function Shuttle() {
   const [confirmModalShow, setConfirmModalShow] = useState(false)
   const [claimModalShow, setClaimModalShow] = useState(false)
   const [value, setValue] = useState('')
-  const [txModalShow, setTxModalShow] = useState(false)
-  const [txModalType, setTxModalType] = useState(TxReceiptModalType.ongoing)
+
   const [txHash, setTxHash] = useState('')
   const {fromChain, toChain, fromTokenAddress, ...others} = queryString.parse(
     location.search,
@@ -230,8 +223,6 @@ function Shuttle() {
           value={value}
           fromToken={fromToken}
           toToken={toToken}
-          setTxModalType={setTxModalType}
-          setTxModalShow={setTxModalShow}
           setTxHash={setTxHash}
           fromAddress={fromAddress}
           toAddress={toAddress}
@@ -254,22 +245,6 @@ function Shuttle() {
           txHash={txHash}
           fromAddress={fromAddress}
           toAddress={toAddress}
-        />
-      )}
-      {txModalShow && (
-        <TransactionReceiptionModal
-          type={txModalType}
-          open={txModalShow}
-          fromChain={fromChain}
-          toChain={toChain}
-          fromToken={fromToken}
-          toToken={toToken}
-          value={value}
-          txHash={txHash}
-          onClose={() => {
-            setTxModalShow(false)
-            setValue('')
-          }}
         />
       )}
     </div>
