+++ src/pages/components/index.js
@@ -17,3 +17,4 @@ export {default as useClaimNotification} from './useClaimNotification'
 export {default as RiskModal} from './RiskModal'
 export {default as ClaimAddressModal} from './ClaimAddressModal'
 export {default as AccountError} from './AccountError'
+export {default as ClaimButton} from './ClaimButton'
